requirements.txt
langchain
python-dotenv
langchain-core
langchain-openai
langchain-ollama
langchain_huggingface
langchain-community
langchain-text-splitters
bs4
faiss-cpu
pypdf